refactor queue constants and namespace
