+++ api/communication-api/src/main/java/com/cognifide/aet/communication/api/queues/WorkerConfig.java
@@ -13,7 +13,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.cognifide.aet.queues;
+package com.cognifide.aet.communication.api.queues;
 
 public interface WorkerConfig {
