+++ osgi-dependencies/selenium/pom.xml
@@ -24,7 +24,7 @@
   <parent>
     <artifactId>osgi-dependencies</artifactId>
     <groupId>com.cognifide.aet</groupId>
-    <version>3.1.1-SNAPSHOT</version>
+    <version>3.2.0-SNAPSHOT</version>
   </parent>
 
   <artifactId>selenium</artifactId>
